refactor: Split javac.ts into focused submodules (phase 1)
Extract lexer, parser, and AST type definitions out of the monolithic
javac.ts into separate files under web/javac/:

- web/javac/ast.ts    — AST node type definitions (ClassDecl, Stmt, Expr …)
- web/javac/lexer.ts  — tokenizer (lex())
- web/javac/parser.ts — recursive-descent parser (parseAll())

web/javac.ts now imports from these submodules and retains only the
code-generation (emit) logic, keeping it focused on a single concern.
